Fixed Lint issues : private fields, shorter declaration, backup, library version upgrade

+++ app/src/main/java/com/github/jmuthu/poovali/model/plant/PlantBatch.java
@@ -126,7 +126,7 @@ public class PlantBatch implements Serializable, DisplayableItem {
     }
 
     public int getProgress() {
-        return (int) getDurationInDays() * 100 / plant.getCropDuration();
+        return getDurationInDays() * 100 / plant.getCropDuration();
     }
 
     public Plant.GrowthStage getStage() {
